+++ src/components/Header/extras/CloseButton.tsx
@@ -1,6 +1,6 @@
 import { MouseEvent } from 'react';
 
-import Icon from '@/components/commons/Icon-new';
+import Icon from '@/components/commons/Icon';
 
 interface CloseButtonProps {
   onClick?: (e?: MouseEvent) => void;
