+++ src/components/Header/extras/ListViewToggleButton.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 import styled from '@emotion/styled';
 
-import Icon from '@/components/commons/Icon-new';
+import Icon from '@/components/commons/Icon';
 
 export type ListViewType = 'grid' | 'list';
