+++ src/components/Header/extras/SaveButton.tsx
@@ -1,6 +1,6 @@
 import { MouseEvent } from 'react';
 
-import Icon from '@/components/commons/Icon-new';
+import Icon from '@/components/commons/Icon';
 
 interface SaveButtonProps {
   onClick?: (e?: MouseEvent) => void;
