+++ src/components/Header/extras/WriteButton.tsx
@@ -1,6 +1,6 @@
 import { MouseEvent } from 'react';
 
-import Icon from '@/components/commons/Icon-new';
+import Icon from '@/components/commons/Icon';
 
 interface WriteButtonProps {
   onClick?: (e?: MouseEvent) => void;
