+++ src/components/search/SearchItem/SearchItem.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useMemo } from 'react';
 import styled from '@emotion/styled';
 
 import useSearchHistory from '@/hooks/useSearchHistory';
-import Icon from '@/components/commons/Icon-new';
+import Icon from '@/components/commons/Icon';
 
 export type SearchIconType = 'search' | 'history';
